requirements.txt
requests>=2.25.0
PyExecJS>=1.5.1
colorama>=0.4.4 